requirements.txt
django
djangorestframework
pandas
lxml
request
bs4